+++ src/MyPackage/Calculate.java
@@ -1,14 +1,14 @@
 package MyPackage;
 
 public class Calculate {
-  static void Add(int num1, int num2) { Add.AddTwoNumbers(num1, num2); }
-  static void Sub(int num1, int num2) {
+  static void Add(float num1, float num2) { Add.AddTwoNumbers(num1, num2); }
+  static void Sub(float num1, float num2) {
     Sub.SubtractTwoNumbers(num1, num2);
   }
-  static void Mul(int num1, int num2) {
+  static void Mul(float num1, float num2) {
     Mul.MultiplyTwoNumbers(num1, num2);
   }
-  static void Div(int num1, int num2) {
+  static void Div(float num1, float num2) {
     Div.DivideTwoNumbers(num1, num2);
   }
 }
