+++ src/MyPackage/Div.java
@@ -1,7 +1,7 @@
 package MyPackage;
 
 public class Div {
-  protected static void DivideTwoNumbers(int num1, int num2) {
+  protected static void DivideTwoNumbers(float num1, float num2) {
     try {
       System.out.println(num1 / num2);
     } catch (ArithmeticException e) {
